+++ controlplane/src/core/bufservices/PlatformService.ts
@@ -179,6 +179,11 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           routingUrl: req.routingUrl,
         });
 
+        await fedGraphRepo.createGraphCryptoKeyPairs({
+          federatedGraphId: federatedGraph.id,
+          organizationId: authContext.organizationId,
+        });
+
         const subgraphs = await subgraphRepo.listByFederatedGraph(req.name, {
           published: true,
         });
@@ -488,6 +493,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           return {
             subgraphs: [],
             graphToken: '',
+            graphRequestToken: '',
             response: {
               code: EnumStatusCode.ERR_NOT_FOUND,
               details: `Federated graph '${req.name}' not found`,
@@ -504,9 +510,27 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
 
         const list = await subgraphRepo.listByFederatedGraph(req.name, { published: true });
 
+        const routerRequestToken = await fedRepo.getGraphSignedToken({
+          federatedGraphId: federatedGraph.id,
+          organizationId: authContext.organizationId,
+        });
+
+        if (!routerRequestToken) {
+          return {
+            subgraphs: [],
+            graphToken: '',
+            graphRequestToken: '',
+            response: {
+              code: EnumStatusCode.ERR,
+              details: 'Router Request token not found',
+            },
+          };
+        }
+
         const tokens = await fedRepo.getRouterTokens({
           organizationId: authContext.organizationId,
           federatedGraphId: federatedGraph.id,
+          limit: 1,
         });
 
         let graphToken: GraphApiKeyDTO;
@@ -550,6 +574,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
             labels: g.labels,
           })),
           graphToken: graphToken.token,
+          graphRequestToken: routerRequestToken,
           response: {
             code: EnumStatusCode.OK,
           },
@@ -1166,6 +1191,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
         const fedGraphRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
         const subgraphRepo = new SubgraphRepository(opts.db, authContext.organizationId);
+        const schemaCheckRepo = new SchemaCheckRepository(opts.db);
 
         const graph = await fedGraphRepo.byName(req.graphName);
 
@@ -1175,7 +1201,10 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               code: EnumStatusCode.ERR_NOT_FOUND,
               details: 'Requested graph does not exist',
             },
+            compositionErrors: [],
+            changes: [],
             affectedGraphs: [],
+            trafficCheckDays: 0,
           };
         }
 
@@ -1188,22 +1217,14 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               code: EnumStatusCode.ERR_NOT_FOUND,
               details: 'Requested check not found',
             },
+            compositionErrors: [],
+            changes: [],
             affectedGraphs: [],
+            trafficCheckDays: 0,
           };
         }
 
-        let addCount = 0;
-        let minusCount = 0;
-        for (const log of checkDetails.changes) {
-          if (log.changeType.includes('REMOVED')) {
-            minusCount += 1;
-          } else if (log.changeType.includes('ADDED')) {
-            addCount += 1;
-          } else if (log.changeType.includes('CHANGED')) {
-            addCount += 1;
-            minusCount += 1;
-          }
-        }
+        const { trafficCheckDays } = await schemaCheckRepo.getFederatedGraphConfigForCheckId(req.checkId, graph.id);
 
         return {
           response: {
@@ -1212,10 +1233,9 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           check,
           affectedGraphs: check.affectedGraphs,
           proposedSubgraphSchemaSDL: check.proposedSubgraphSchemaSDL,
-          changeCounts: {
-            additions: addCount,
-            deletions: minusCount,
-          },
+          changes: checkDetails.changes,
+          compositionErrors: checkDetails.compositionErrors,
+          trafficCheckDays,
         };
       });
     },
@@ -3553,6 +3573,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const tokens = await fedRepo.getRouterTokens({
           organizationId: authContext.organizationId,
           federatedGraphId: federatedGraph.id,
+          limit: 100,
         });
 
         return {