+++ studio/src/components/layout/graph-layout.tsx
@@ -11,48 +11,44 @@ import {
 } from "@radix-ui/react-icons";
 import { useQuery } from "@tanstack/react-query";
 import { useRouter } from "next/router";
-import { getFederatedGraphByName } from "@wundergraph/cosmo-connect/dist/platform/v1/platform-PlatformService_connectquery";
+import {
+  getFederatedGraphByName,
+  getFederatedGraphs,
+} from "@wundergraph/cosmo-connect/dist/platform/v1/platform-PlatformService_connectquery";
 import { GetFederatedGraphByNameResponse } from "@wundergraph/cosmo-connect/dist/platform/v1/platform_pb";
 import { ReactNode, createContext, useContext, useMemo } from "react";
-import { PiGitBranch } from "react-icons/pi";
+import { PiGitBranch, PiDevices } from "react-icons/pi";
 import { EmptyState } from "../empty-state";
 import { Button } from "../ui/button";
 import { Loader } from "../ui/loader";
 import { LayoutProps } from "./layout";
-import { Nav, NavLink } from "./nav";
 import { EnumStatusCode } from "@wundergraph/cosmo-connect/dist/common/common_pb";
-import { addDays, formatDistance } from "date-fns";
-import { UserContext } from "../app-provider";
-import { calURL } from "@/lib/constants";
-import Link from "next/link";
-import { MdDevices } from "react-icons/md";
-
-const icons: { [key: string]: ReactNode } = {
-  Overview: <HomeIcon />,
-  Subgraphs: <Component2Icon />,
-  Explorer: <PlayIcon />,
-  Schema: <FileTextIcon />,
-  Changelog: <PiGitBranch />,
-  Checks: <CheckCircledIcon />,
-  Analytics: <ChartBarIcon className="h-4 w-4" />,
-  Clients: <MdDevices className="h-4 w-4" />,
-};
+import { SideNav, NavLink } from "./sidenav";
+import { PageHeader } from "./head";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "../ui/select";
+import { cn } from "@/lib/utils";
 
 export interface GraphContextProps {
   graph: GetFederatedGraphByNameResponse["graph"];
   subgraphs: GetFederatedGraphByNameResponse["subgraphs"];
   graphToken: string;
+  graphRequestToken: string;
 }
 
 export const GraphContext = createContext<GraphContextProps | undefined>(
   undefined,
 );
 
-const GraphLayout = ({ children }: LayoutProps) => {
+export const GraphLayout = ({ children }: LayoutProps) => {
   const router = useRouter();
   const organizationSlug = router.query.organizationSlug as string;
   const slug = router.query.slug as string;
-  const user = useContext(UserContext);
 
   const { data, isLoading, error, refetch } = useQuery(
     getFederatedGraphByName.useQuery({
@@ -68,6 +64,7 @@ const GraphLayout = ({ children }: LayoutProps) => {
       graph: data.graph,
       subgraphs: data.subgraphs,
       graphToken: data.graphToken,
+      graphRequestToken: data.graphRequestToken,
     };
   }, [data]);
 
@@ -75,27 +72,31 @@ const GraphLayout = ({ children }: LayoutProps) => {
     const basePath = `/${organizationSlug}/graph/${slug}`;
 
     return [
-      { title: "Overview", href: basePath, icon: <HomeIcon /> },
+      {
+        title: "Overview",
+        href: basePath,
+        icon: <HomeIcon className="h-4 w-4" />,
+      },
       {
         title: "Subgraphs",
         href: basePath + "/subgraphs",
-        icon: <Component2Icon />,
+        icon: <Component2Icon className="h-4 w-4" />,
       },
       {
         title: "Playground",
         href: basePath + "/playground",
-        icon: <PlayIcon />,
+        icon: <PlayIcon className="h-4 w-4" />,
       },
       {
         title: "Schema",
         href: basePath + "/schema",
         matchExact: false,
-        icon: <FileTextIcon />,
+        icon: <FileTextIcon className="h-4 w-4" />,
       },
       {
         title: "Clients",
         href: basePath + "/clients",
-        icon: <MdDevices className="h-4 w-4" />,
+        icon: <PiDevices className="h-4 w-4" />,
       },
       {
         title: "Analytics",
@@ -106,13 +107,13 @@ const GraphLayout = ({ children }: LayoutProps) => {
       {
         title: "Changelog",
         href: basePath + "/changelog",
-        icon: <PiGitBranch />,
+        icon: <PiGitBranch className="h-4 w-4" />,
       },
       {
         title: "Checks",
         href: basePath + "/checks",
         matchExact: false,
-        icon: <CheckCircledIcon />,
+        icon: <CheckCircledIcon className="h-4 w-4" />,
       },
     ];
   }, [slug, organizationSlug]);
@@ -144,13 +145,110 @@ const GraphLayout = ({ children }: LayoutProps) => {
 
   return (
     <div className="2xl:flex 2xl:flex-1 2xl:flex-col 2xl:items-center">
-      <div className="min-h-screen w-full max-w-screen-4xl bg-background font-sans antialiased">
-        <Nav links={links}>{render}</Nav>
+      <div className="flex min-h-screen w-full flex-1 flex-col bg-background font-sans antialiased lg:grid lg:grid-cols-[auto_1fr] lg:divide-x">
+        <SideNav links={links} />
+        <main className="flex-1">{render}</main>
       </div>
     </div>
   );
 };
 
-export const getGraphLayout = (page: React.ReactNode) => {
-  return <GraphLayout>{page}</GraphLayout>;
+export const GraphSelect = () => {
+  const { data } = useQuery(getFederatedGraphs.useQuery());
+
+  const router = useRouter();
+  const slug = router.query.slug as string;
+  const organizationSlug = router.query.organizationSlug as string;
+  if (router.pathname.split("/")[2] !== "graph") return null;
+
+  return (
+    <Select
+      value={slug}
+      onValueChange={(gID) => router.push(`/${organizationSlug}/graph/${gID}`)}
+    >
+      <SelectTrigger
+        value={slug}
+        className="flex h-8 w-auto gap-x-2 border-0 bg-transparent pl-3 pr-1 text-muted-foreground shadow-none data-[state=open]:bg-accent data-[state=open]:text-accent-foreground hover:bg-accent hover:text-accent-foreground focus:ring-0"
+      >
+        <SelectValue aria-label={slug}>{slug}</SelectValue>
+      </SelectTrigger>
+      <SelectContent className="min-w-[200px]">
+        {data?.graphs?.map(({ name }) => {
+          return (
+            <SelectItem key={name} value={name}>
+              {name}
+            </SelectItem>
+          );
+        })}
+      </SelectContent>
+    </Select>
+  );
+};
+
+export const getGraphLayout = (page: React.ReactNode, { title }: any = {}) => {
+  return (
+    <GraphLayout>
+      <PageHeader title={`${title} | Studio`}>{page}</PageHeader>
+    </GraphLayout>
+  );
+};
+
+export interface TitleLayoutProps {
+  breadcrumbs?: React.ReactNode[];
+  title: React.ReactNode;
+  subtitle: string;
+  items?: React.ReactNode;
+  toolbar?: React.ReactNode;
+  noPadding?: boolean;
+  children?: React.ReactNode;
+}
+
+export const GraphPageLayout = ({
+  title,
+  breadcrumbs,
+  items,
+  toolbar,
+  noPadding,
+  children,
+}: TitleLayoutProps) => {
+  const breadcrumb = (
+    <div className="-ml-2 flex flex-row items-center space-x-2 text-sm">
+      <GraphSelect /> <span className="text-muted-foreground">/</span>
+      {breadcrumbs?.map((b) => (
+        <>
+          <span className="text-muted-foreground hover:text-current">{b}</span>
+          <span className="text-muted-foreground">/</span>
+        </>
+      ))}
+      <h1 className="truncate whitespace-nowrap font-medium">{title}</h1>
+    </div>
+  );
+
+  return (
+    <div className="flex flex-col lg:h-screen">
+      <div className="bg-background">
+        <div
+          className={cn(
+            "flex flex-col justify-between gap-y-4 px-4 pb-2 pt-4 lg:flex-row lg:items-center lg:px-8",
+            {
+              "border-b": !toolbar,
+              "pb-4": !toolbar,
+            },
+          )}
+        >
+          {breadcrumb}
+          {items}
+        </div>
+        {toolbar}
+      </div>
+      <div
+        className={cn(
+          "h-auto flex-1 overflow-y-auto",
+          noPadding !== true && "px-4 py-6 lg:px-8",
+        )}
+      >
+        {children}
+      </div>
+    </div>
+  );
 };