+++ studio/src/pages/[organizationSlug]/index.tsx
@@ -1,3 +1,4 @@
+import { Loader } from "@/components/ui/loader";
 import { NextPageWithLayout } from "@/lib/page";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
@@ -9,17 +10,23 @@ const DashboardPage: NextPageWithLayout = () => {
   useEffect(() => {
     if (!organizationSlug) return;
     const url = new URL(
-      window.location.origin + router.basePath + router.asPath
+      window.location.origin + router.basePath + router.asPath,
     );
     const params = new URLSearchParams(url.search);
     router.replace(
       params.size !== 0
         ? `/${organizationSlug}/graphs?${params}`
-        : `/${organizationSlug}/graphs`
+        : `/${organizationSlug}/graphs`,
     );
   }, [router, organizationSlug]);
 
-  return null;
+  return (
+    <div className="fixed inset-0 flex items-center justify-center bg-background">
+      <Loader />
+    </div>
+  );
 };
+
+DashboardPage.getLayout = (page) => page;
 
 export default DashboardPage;